pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "simcortexpp"
version = "0.1.0"
description = "SimCortexPP (SCPP) pipeline stages for Nipoppy (preproc + segmentation)"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Your Name" }]

dependencies = [
  "nipoppy>=0.4.5",
  "typer>=0.12",
  "numpy>=1.24",
  "pandas>=2.0",
  "nibabel>=5.2",
  "hydra-core>=1.3",
  "omegaconf>=2.3",
  "tqdm>=4.66",
  "tensorboard>=2.14",
]

[project.optional-dependencies]
seg = [
  "monai>=1.3",
  # Torch is often installed explicitly on HPC; keep it optional.
  # If you want pip to install it too, uncomment:
  # "torch>=2.0"
]
preproc = [
  "trimesh>=4.0",
]

[project.scripts]
scpp = "simcortexpp.cli.main:app"

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
simcortexpp = [
  "configs/**/*.yaml",
]
